+++ x/fswap/client/proposal_handler.go
@@ -1,0 +1,8 @@
+package client
+
+import (
+	"github.com/Finschia/finschia-sdk/x/fswap/client/cli"
+	govclient "github.com/Finschia/finschia-sdk/x/gov/client"
+)
+
+var ProposalHandler = govclient.NewProposalHandler(cli.NewCmdMakeSwapProposal)